Moved Pageable from a module to a class Allows Requestable to delegate the question to Pageable More clearly defines what Pageable needs to run
